backend/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
email-validator==2.3.0
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
psycopg[binary]==3.3.3
python-dotenv==1.2.2
